readock/requirements.txt
jinja2
pyyaml